refactor(changelog): Move parser out of releaser
